+++ src/reportingApi.ts
@@ -1,4 +1,5 @@
 import * as vitest from 'vitest';
+import clientHelpers from '@reportportal/client-javascript/lib/helpers';
 import * as Models from './models';
 import { isRPTaskMeta } from './utils';
 
@@ -12,6 +13,7 @@ const injectRPTaskMeta = (task: vitest.Task) => {
     rpMeta: {
       test: {
         logs: [],
+        attributes: [],
       },
     },
   };
@@ -21,16 +23,42 @@ const attachment = (task: vitest.Task, data: Models.Attachment, description?: st
   injectRPTaskMeta(task);
   (task.meta as Models.RPTaskMeta).rpMeta.test.logs.push({
     file: data,
-    time: Date.now(),
+    time: clientHelpers.now(),
     message: description || data.name,
   });
 };
 
+const attributes = (task: vitest.Task, data: Models.Attribute[]) => {
+  injectRPTaskMeta(task);
+  const rpMeta = (task.meta as Models.RPTaskMeta).rpMeta;
+  rpMeta.test.attributes = [...rpMeta.test.attributes, ...data];
+};
+
+const testCaseId = (task: vitest.Task, data: string) => {
+  injectRPTaskMeta(task);
+  (task.meta as Models.RPTaskMeta).rpMeta.test.testCaseId = data;
+};
+
+const description = (task: vitest.Task, data: string) => {
+  injectRPTaskMeta(task);
+  const rpMeta = (task.meta as Models.RPTaskMeta).rpMeta;
+  if (rpMeta.test.description) {
+    rpMeta.test.description = `${rpMeta.test.description}\n${data}`;
+  } else {
+    rpMeta.test.description = data;
+  }
+};
+
 export const ReportingApi: Models.ReportingApi = {
   attachment,
+  attributes,
+  testCaseId,
+  description,
 };
 
 export const bindReportingApi = (task: vitest.Task): Models.GlobalReportingApi => ({
-  attachment: (data: Models.Attachment, description?: string) =>
-    attachment(task, data, description),
+  attachment: (data: Models.Attachment, message?: string) => attachment(task, data, message),
+  attributes: (data: Models.Attribute[]) => attributes(task, data),
+  testCaseId: (data: string) => testCaseId(task, data),
+  description: (data: string) => description(task, data),
 });