+++ src/types/@reportportal/client-javascript/lib/helpers.d.ts
@@ -15,16 +15,7 @@
  *
  */
 
-declare module '@reportportal/client-javascript' {
-  export default class {
-    constructor(config: any, agentInfo?: any);
-
-    public startLaunch(launchObj: any): any;
-    public finishLaunch(launchId: string, launchObj: any): any;
-    public startTestItem(itemObj: any, launchId: string, parentId?: string): any;
-    public finishTestItem(itemId: string, itemObj: any): any;
-    public sendLog(itemId: string, itemObj: any, fileObj?: any): any;
-    public checkConnect(): any;
-    public helpers: any;
-  }
+declare module '@reportportal/client-javascript/lib/helpers' {
+  export function now(): string;
+  export function formatMicrosecondsToISOString(microseconds: number): string;
 }