+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ### Added
-- Test case id, attributes and description providing via Reporting API.
+- Test case id, attributes and description providing via [Reporting API](./README.md#reporting-api-methods).
 ### Changed
-- `@reportportal/client-javascript` bumped to version `5.1.4`, new `launchUuidPrintOutput` types introduced: 'FILE', 'ENVIRONMENT'.
+- The agent now supports reporting the time for launches, test items and logs with microsecond precision in the ISO string format.
+  For logs, microsecond precision is available on the UI from ReportPortal version 24.2.
+- `@reportportal/client-javascript` bumped to version `5.3.0`, new `launchUuidPrintOutput` types introduced: 'FILE', 'ENVIRONMENT'.
 ### Security
 - Updated versions of vulnerable packages (braces).